* commit '0.8.1':
  0.8.1 release.
  FIX: fix segfault in cases of infeasible nu (NuSVM)
  Test for non-contiguous input for svms
  COSMIT: Better fix for continuity in SVMs.
  BUG: Make SVMs work on non contiguous arrays
  minor -- pass verbose into LARS in the test case
  test case for previous commit
  FIX: lars_path -- assure that at least some features get added if necessary
  DOC: minor spellings fixes in pls.py
  FIX: strings are not necessarily singletones + catch mistakes earlier
  FIX: reflect SVC API change (eps -> tol) in doc/tutorial.rst

+++ scikits/learn/svm/src/libsvm/svm.cpp
@@ -3038,7 +3038,7 @@ const char *PREFIX(check_parameter)(const PREFIX(problem) *prob, const svm_param
 				{
 					max_nr_class *= 2;
 					label = (int *)realloc(label,max_nr_class*sizeof(int));
-					count = (double *)realloc(count,max_nr_class*sizeof(int));
+					count = (double *)realloc(count,max_nr_class*sizeof(double));
 
 				}
 				label[nr_class] = this_label;
